% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SinONTbar.R
\name{ONTBarcodeSeq}
\alias{ONTBarcodeSeq}
\title{extract barcode seq form ONT data according to 1 pattern}
\usage{
ONTBarcodeSeq(
  read,
  qual,
  patta,
  patta2,
  StartMismatch = 5,
  BarcodeLength = 8,
  umilength = 12,
  linkerlength = 16,
  barcode_max_dis = 12
)
}
\arguments{
\item{read}{fastq read in DNAstring format}

\item{qual}{fastq quality}

\item{patta}{linker1}

\item{patta2}{linker2}

\item{StartMismatch}{the mismatch base number threshold when searching for pattern in ONT reads using matchPattern}

\item{BarcodeLength}{Singleron  single barcode length}

\item{umilength}{Singleron  umi length}

\item{linkerlength}{Singleron single linker length}

\item{barcode_max_dis}{the maximum distance between two linkers, if there is no indels or delete, the distance is BarcodeLength 8.}
}
\value{
Barcode + umi sequence in DNAstringset format extracted from ONT data.
}
\description{
extract barcode seq form ONT data according to 1 pattern
}
