% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SinONTbar.R
\name{PattafourSeq}
\alias{PattafourSeq}
\title{extract barcode seq form ONT data according to 4 pattern}
\usage{
PattafourSeq(ONTfastq)
}
\arguments{
\item{ONTfastq}{ONTfastq site, for example: /mnt/data1/zhangdan/Proj/Mouse_iso_atlas/scnapbar/sc_NGScDNA_C57_F_2m_c-kit+/data/nanopore_small.fq}
}
\value{
ONT barcode sequence
}
\description{
extract barcode seq form ONT data according to 4 pattern
}
