% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SinONTbar.R
\name{SequenceSimilarity}
\alias{SequenceSimilarity}
\title{Calculate two sequence similarity}
\usage{
SequenceSimilarity(x)
}
\arguments{
\item{x}{please check parSeqSim input}
}
\value{

}
\description{
Calculate two sequence similarity
}
